feat(generative): improve the messages for the `ViewWorkflow` service
In the case of the container types `WorkflowParameterType::ARRAY` and
`WorkflowParameterType::OBJECT` the types of their elements is not
provided.

I tried finding a way to make this work but the `pydantic` validators
really did not like the recursion required for this.

This information can still be communicated though by the default and
example values.